created blog_test_suite and user_test_suite
